requirements.txt
tensorflow 
keras 
opencv-python 
matplotlib 
numpy 
pandas 
scikit-learn 
flask
gradio
pillow
streamlit
plotly
reportlab